pyproject.toml
[project]
name = "redoing-my-first-python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.1.3",
    "pandas-datareader>=0.10.0",
    "plotly>=5.24.1",
    "streamlit>=1.40.0",
    "yfinance==0.1.67",
]
